* renamed call_run to _call_run to make clear that it is not part of the gt4py user interface. * accepting numpy integer types for domain, origin in call and storage interfaces
